Seungwookim (#4)
* expo-sqlite installed

* split to presenter and container

* ORM model added

* async error handled in useEffect

* WIP

* 주석 해제

* table name changed

* PieChart WIP

* volume calculator WIP

* 통계 렌더링 WIP

* 통계 LineChart WIP

* Pie chart done

* changeUnit function added

* changeUnit function added

* Report page done

* setLoading 주석 제거

+++ model/Set.js
@@ -17,7 +17,7 @@ export default class Set extends BaseModel {
   static get columnMapping() {
     return {
       id: { type: types.INTEGER, primary_key: true },
-      weight: { type: types.INTEGER },
+      weight: { type: types.FLOAT },
       rep: { type: types.INTEGER },
       time: { type: types.INTEGER },
     }
